+++ sph+nln/eos.h
@@ -5,10 +5,10 @@ extern double eos_u;
 #ifndef EOS_H
 #define EOS_H
 typedef struct material_s {
-	float rho0;
-	float A, B, a, b, alpha, beta;
-	float Eiv, Ecv;
-	float u0, umelt;
-	float mu, yield, pweib, cweib;
+	double rho0;
+	double A, B, a, b, alpha, beta;
+	double Eiv, Ecv;
+	double u0, umelt;
+	double mu, yield, pweib, cweib;
 } Material_t;
 #endif
