+++ sph+nln/physics_sph.h
@@ -589,7 +589,7 @@ double liquid_eos (double k_bulk, double eta);
 double murnaghan_eos(double k_bulk, double n_M, double eta);
 void setconst1(Material_t *m);
 void setconst2(Material_t *m);
-void tillotson_eos (float rho, float u, Material_t *m, float *pressure, float *cs);
+void tillotson_eos (double rho, double u, Material_t *m, double *pressure, double *cs);
 double anton_schmidt_eos(double k_bulk, double power_n, double eta);
 
 /* In newtraph.c */
